Add method baca file csv di handler makanan dll
Tambah folder assets dan file csv di dalamnya, update atribut Makanan
"kalori" dari double ke int

+++ Project/src/model/Makanan.java
@@ -3,7 +3,7 @@ package model;
 public class Makanan {
 
 	private String nama;
-	private double kalori;
+	private int kalori;
 	private double protein;
 	private double karbohidrat;
 	private double lemak;
